SCRUM-68 Revised login code, Added clearUserDB()
Website should properly include a username AND password based login.
There is no way for the user to explicitly choose whether to sign up/in yet, that will be addressed in a later task.

However, the backend implementation for sign up/in and login is complete.

Additionally, there is now a clearUserDB() function, callable via the console. It deletes everything from the UserDB. This is useful during development if there are duplicate users.